requirements.txt
attrs==17.4.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
idna==2.6
requests==2.18.4
requests-cache==0.4.13
urllib3==1.22
urltools==0.3.2
